refactor: Application 구현
- 기존의 Controller에서 전체 로직을 수행하는것을 application 클래스에서 수행하도록 수정함.
